feat(schema): route source-material capture through sensd (#76)
testing and ergonomics:
- add property-test suites for path sanitization, time ranges, and validation invariants, plus new preludes for gateway, ingestd, the satellite SDK, and services, so the schema/runtime split can be exercised and imported without deep path churn
- flatten `sinex-core` exports and document the import style, which reduces the amount of local facade code needed in tests and satellites while the schema crate is introduced

schema and crate ownership:
- fold more operations-log, provenance, and migration fixups into the base schema, then rename `sinex-migrations` to `sinex-schema` and move shared schema definitions and record types there, giving migrations and table metadata a dedicated crate instead of living beside runtime code
- move typed ids back into `sinex-core` after the first split, keeping runtime-facing ids and model aliases on the core crate while leaving schema utilities and migration helpers in `sinex-schema`
- align `core_events`, `source_materials`, and sensor-job schema definitions with the runtime code that now depends on them, including the stricter provenance constraint for material-backed events

source-material and ingest flow:
- route blob lifecycle events through ingestd, add namespaced environment/resource helpers, and stop writing some state or lock events directly, so runtime subsystems use the same ingestion boundary instead of bypassing it
- add `sinex-sensd` gRPC service and proto definitions, then align source-material and sensor-job schema around that service boundary
- route terminal, desktop, and document acquisition through source materials, making `sensd` the shared acquisition path instead of a side subsystem

canonicalization:
- require `anchor_byte` for material-backed terminal events, preserving exact provenance for replay and material slicing
- expand `sinex-terminal-command-canonicalizer` into a larger automaton and add schema constants used by repository callers, so terminal canonicalization can consume the richer source-material-backed event model

+++ crate/lib/sinex-core/src/db/repositories/knowledge_graph.rs
@@ -601,7 +601,7 @@ impl<'a> KnowledgeGraphRepository<'a> {
             *relation.from_entity_id.as_ulid() as _,
             *relation.to_entity_id.as_ulid() as _,
             relation.relation_type,
-            strength,
+            strength as f32,
             metadata,
             valid_from,
             relation.valid_until,
